checkout: Generate a new idempotency key after cart restore (#261)
* checkout: Generate a new idempotency key after cart restore

When a payment is aborted/canceled the checkout controller regenerates a new idempotency key. If an early place happened the cart at this stage is potentially not accessible any more. We therefore need to first cancel the order / restore the cart before setting the new PaymentSelection / Idempotency Key.

This PR changes the orders of the calls.

* checkout: fix ci