checkout: Add CartValidationResult to place order context if in failed state (#319)
In the GraphQL implementation we expose the cart validation result if the ValidateCart state fails. This information is missing in the JSON API so that you only see that the cart is invalid but not which rule was triggered. This PR therefore adds an additional field to the place order context when it's in failed state due to an invalid cart.